requirements.txt
streamlit
python-dotenv

langchain-core
langchain-community
langchain-text-splitters
langchain-groq

faiss-cpu
fastembed

numpy
